refactor strg_fifo for memtilebuilder
